🎨 improve code, refactoring lib import/export
