+++ components/Layout/index.tsx
@@ -8,15 +8,13 @@ const Footer = dynamic(() => import("components/Footer"))
 import { ISEO } from "interfaces"
 import { Main } from "styles/shared"
 
-const Layout = ({ children, seo }: { children: ReactNode; seo: ISEO }) => {
-  return (
-    <>
-      <Seo seo={seo} />
-      <Nav />
-      <Main>{children}</Main>
-      <Footer />
-    </>
-  )
-}
+const Layout = ({ children, seo }: { children: ReactNode; seo: ISEO }) => (
+  <>
+    <Seo seo={seo} />
+    <Nav />
+    <Main>{children}</Main>
+    <Footer />
+  </>
+)
 
 export default Layout
