+++ components/Wysiwyg/index.tsx
@@ -1,8 +1,8 @@
 import rehypeRaw from "rehype-raw"
 import * as Styled from "./styled"
 
-const Wysiwyg = ({ content }: any) => {
-  return <Styled.Wysiwyg rehypePlugins={[rehypeRaw]}>{content}</Styled.Wysiwyg>
-}
+const Wysiwyg = ({ content }: any) => (
+  <Styled.Wysiwyg rehypePlugins={[rehypeRaw]}>{content}</Styled.Wysiwyg>
+)
 
 export default Wysiwyg
