+++ lib/index.ts
@@ -1,0 +1,4 @@
+export { getStrapiURL, fetchAPI } from "./api"
+export { getPaths } from "./getPaths"
+export { GlobalContext } from "./global"
+export { getStrapiMedia } from "./media"
